adding negative test cases and refactoring code

+++ src/main/java/nagarro/service/impl/PersonServiceImpl.java
@@ -3,6 +3,7 @@ package nagarro.service.impl;
 import nagarro.dao.PersonDAO;
 import nagarro.dto.PersonDTO;
 import nagarro.entity.Person;
+import nagarro.exception.PersonNotFoundException;
 import nagarro.service.PersonService;
 import org.jvnet.hk2.annotations.Service;
 
@@ -14,49 +15,61 @@ import java.util.stream.Collectors;
 public class PersonServiceImpl implements PersonService {
     private final PersonDAO personDAO;
 
-    public PersonServiceImpl(PersonDAO personDAO){
+    public PersonServiceImpl(final PersonDAO personDAO) {
         this.personDAO = personDAO;
     }
 
     @Override
-    public final PersonDTO insertPerson(PersonDTO personDTO){
-        var person = new Person(0, personDTO.getName(), personDTO.getAge());
-        int generatedId = personDAO.insertPerson(person);
+    public final PersonDTO createPerson(final PersonDTO personDTO) {
+        validatePersonDTO(personDTO);
+        final var person = new Person(0, personDTO.getName(), personDTO.getAge());
+        final int generatedId = personDAO.createPerson(person);
         person.setId(generatedId);
         return convertPersonToPersonDTO(person);
     }
 
     @Override
-    public final List<PersonDTO> getAllPersons(){
-        return personDAO.getAllPersons().stream()
-                .map(person -> new PersonDTO(person.getId(), person.getName(), person.getAge()))
-                .collect(Collectors.toList());
+    public final List<PersonDTO> getAllPersons() {
+        return personDAO.getAllPersons().stream().map(person -> new PersonDTO(person.getId(), person.getName(), person.getAge())).collect(Collectors.toList());
     }
 
     @Override
-    public final PersonDTO getPersonById(int id){
-        final Optional<Person> person = Optional.ofNullable(personDAO.getPersonById(id));
-        if(person.isPresent()){
+    public final PersonDTO getPersonById(final int id) throws PersonNotFoundException {
+        Optional<Person> person = personDAO.getPersonById(id);
+        if (person.isPresent()) {
             return convertPersonToPersonDTO(person.get());
-        }else{
-            return null;
+        } else {
+            throw new PersonNotFoundException("Person with id " + id + " not found");
         }
     }
 
     @Override
-    public final void updatePerson(PersonDTO personDTO){
-        final var person = new Person(personDTO.getId(), personDTO.getName(), personDTO.getAge());
-        System.out.println(person.getAge()+" "+person.getName()+" "+person.getId());
+    public final void updatePerson(final int id, final PersonDTO personDTO) {
+        validatePersonDTO(personDTO);
+        if (personDAO.getPersonById(id).isEmpty()) {
+            throw new PersonNotFoundException("Person with id " + id + " not found");
+        }
+        final var person = new Person(id, personDTO.getName(), personDTO.getAge());
         personDAO.updatePerson(person);
     }
 
     @Override
-    public final void deletePerson(int id){
-        personDAO.deletePerson(id);
+    public final void deletePerson(final int id) throws PersonNotFoundException {
+        Optional<Person> personToDelete = personDAO.getPersonById(id);
+        if (personToDelete.isPresent()) {
+            personDAO.deletePerson(id);
+        } else {
+            throw new PersonNotFoundException("Person with id " + id + " not found");
+        }
     }
 
-    public final PersonDTO convertPersonToPersonDTO(Person person){
-        final var personDTO = new PersonDTO(person.getId(), person.getName(), person.getAge());
-        return personDTO;
+    private final PersonDTO convertPersonToPersonDTO(final Person person) {
+        return new PersonDTO(person.getId(), person.getName(), person.getAge());
+    }
+
+    private void validatePersonDTO(final PersonDTO personDTO) {
+        if (personDTO.getName() == null || personDTO.getName().isEmpty() || personDTO.getAge() < 0) {
+            throw new IllegalArgumentException("Invalid person data");
+        }
     }
 }
